fix: resolve Rust 1.95 clippy lint drift across 17 files
Fixes ~60 pre-existing clippy errors introduced by new lints in Rust 1.95:
- needless_range_loop: convert index loops to iterator patterns
- manual_clamp: replace .max().min() with .clamp()
- too_many_arguments: add #[allow] with reason on functions with >7 args
- missing_safety_doc: add # Safety sections to unsafe functions
- manual_memcpy: add #[allow] with reason on packed struct field copies
- doc_lazy_continuation: add blank lines before Total: in doc comments
- type_complexity: add #[allow] with reason on complex type alias
- sliced_string_as_bytes: slice bytes before calling .to_vec()
- approx_constant: replace 3.14 with std::f32::consts::PI
- assertions_on_constants: remove assert!(true)
- explicit_counter_loop: use .enumerate() or direct iteration
- unused_mut / unused_variables: remove unused bindings

All changes are mechanical; no functional changes.

+++ src/cpu/kernels/q6.rs
@@ -78,18 +78,18 @@ impl BlockQ6K {
                 let sc3 = d * (self.scales[sc_base + is + 4] as f32);
                 let sc4 = d * (self.scales[sc_base + is + 6] as f32);
 
-                let q1 = ((self.ql[ql_base + l] & 0xF) as u8 | ((self.qh[qh_base + l] & 3) << 4))
+                let q1 = ((self.ql[ql_base + l] & 0xF) | ((self.qh[qh_base + l] & 3) << 4))
                     as i8 as f32
                     - 32.0;
-                let q2 = ((self.ql[ql_base + l + 32] & 0xF) as u8
+                let q2 = ((self.ql[ql_base + l + 32] & 0xF)
                     | (((self.qh[qh_base + l] >> 2) & 3) << 4)) as i8
                     as f32
                     - 32.0;
-                let q3 = (((self.ql[ql_base + l] >> 4) & 0xF) as u8
+                let q3 = (((self.ql[ql_base + l] >> 4) & 0xF)
                     | (((self.qh[qh_base + l] >> 4) & 3) << 4)) as i8
                     as f32
                     - 32.0;
-                let q4 = (((self.ql[ql_base + l + 32] >> 4) & 0xF) as u8
+                let q4 = (((self.ql[ql_base + l + 32] >> 4) & 0xF)
                     | (((self.qh[qh_base + l] >> 6) & 3) << 4)) as i8
                     as f32
                     - 32.0;
